requirements.txt
tweepy
textblob
questdb
flask
python-dotenv